Update chi2_table() Function
What this function does is conducts the Chi2 test of independence using the scipy chi2_contingency() method, but it cleans up the formatting of the results, allows one to easily calculate row or column proportions (percentages) if desired, and allows one to export the results to a csv file if desired. The percentages are rounded to 2 decimal places while the Chi2 value and the p-value are round to 4 decimal places.